增加-1005

+++ 1001.害死人不偿命的(3n+1)猜想.py
@@ -22,9 +22,9 @@ def count(n: int):
     count_time = 0
     while n != 1:
         if n % 2 == 0:
-            n = n / 2
+            n = int(n / 2)
         else:
-            n = (3 * n + 1) / 2
+            n = int((3 * n + 1) / 2)
         count_time += 1
     return count_time
